Added 4 additional Lanczos filter variants (#43)
* New Lanczos filter variants

Using the usual kernels for N=2 and N=3. 4, 8 and 16 phases normalized for 128 and 256 (10bit). Meant to reduce horizontal blur on N64 core when VI bilinear is in use. Less phases produce a closer to original output while adding some diffusion.

* Adding additional Lanczos filters

Using kernels for N=2 and N=3, 256 phases normalized for 128 and 256 (10bit). Realized these filters might be useful for other cores besides N64. Additionally because of the fine granularity these are useful for many screen resolutions