N64 specific filters (#44)
* New Lanczos filter variants

Using the usual kernels for N=2 and N=3. 4, 8 and 16 phases normalized for 128 and 256 (10bit). Meant to reduce horizontal blur on N64 core when VI bilinear is in use. Less phases produce a closer to original output while adding some diffusion.

* Adding additional Lanczos filters

Using kernels for N=2 and N=3, 256 phases normalized for 128 and 256 (10bit). Realized these filters might be useful for other cores besides N64. Additionally because of the fine granularity these are useful for many screen resolutions

* Two new N64 specific filters

These filters are meant to be used horizontally only. One reduces blur when VI bilinear is ON, the other one adds a tiny bit of blur when VI bilinear is OFF